requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
firebase-admin==6.2.0
python-dotenv==1.0.0
pyjwt==2.8.0
requests==2.31.0
psycopg2-binary==2.9.7
redis==4.6.0
celery==5.3.1
twilio==8.5.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==1.5.3
schedule==1.2.0
gunicorn==21.2.0